Added cleanup steps

+++ generate-character-card.sh
@@ -16,16 +16,26 @@ if [ ! -f "in/markdown/$1.md" ]; then
 fi
 
 # Pass the character's markdown file to the python script to convert it to yaml.
-python rpg-cards-py-obsidian-md-to-yaml/convert_character.py "in/markdown/$1.md" "./out/yaml"
+python3 rpg-cards-py-obsidian-md-to-yaml/convert_character.py "in/markdown/$1.md" "in"
 
-# Copy the character's yaml file to the Typst input location and overwrite the file there, if present.
-cp "out/yaml/$1.yaml" "rpg-cards-typst-templates/in/character.yaml"
+# Check if the character's yaml file exists.
+if [ ! -f "in/$1.yaml" ]; then
+  echo "Character not found."
+  exit 1
+fi
 
 # Pull the name of the character's image file from their yaml file.
-image=$(grep "image:" "out/yaml/$1.yaml" | cut -d ' ' -f 2)
+image=$(grep "image:" "in/$1.yaml" | cut -d ' ' -f 2)
+
+# Copy the character's yaml file to the Typst input location and overwrite the file there, if present.
+mv "in/$1.yaml" rpg-cards-typst-templates/in/character.yaml
 
 # Copy the character's image to the Typst input location and overwrite the file there, if present.
 cp "in/images/$image" "rpg-cards-typst-templates/in/$image"
 
 # Run the character card generation script.
-typst compile rpg-cards-typst-templates/src/cards/character/landscape.typ "out/pdf/$1.pdf" --root rpg-cards-typst-templates
+typst compile rpg-cards-typst-templates/src/cards/character/landscape.typ "out/$1.pdf" --root rpg-cards-typst-templates
+
+# Clean up the Typst input location.
+rm "rpg-cards-typst-templates/in/character.yaml"
+rm "rpg-cards-typst-templates/in/$image"
